refactor: extract PipelineBuilderExceptionTests nested doubles to TestDoubles (#306)
Moves the exception-scenario queries/handlers (each Result with its
owning query), TestExceptionDecorator<,>, and DecoratorExceptionAttribute
verbatim into TestDoubles/. Private nested types promoted to internal
(attribute stays public, as before). The test references them via a
using; Moq factory/registry usage unchanged (swapped in Task 9).

Satisfies FR4 (extraction half), FR5, AC6, AC7. Structural; suite green
at baseline.
